blip-0033: commit fees channel flag
Standard lightning channels require the channel opener to pay the
mining fees for the commitment transaction and mutual close transaction.
This requirement forces the channel opener to contribute to the funding
transaction and makes it impossible to open channels by only purchasing
inbound liquidity from a remote node.

We want to allow nodes to opt into paying the commitment transaction
fees even though they're not the channel opener.