blip-0036: on-the-fly channel funding
Payments sent to mobile wallets often fail because the recipient doesn't
have enough inbound liquidity to receive it. we add a mechanism to
create an on-chain transaction on-the-fly before relaying such payments,
which allow them to be relayed once the on-chain transaction is accepted
by both peers.

This protocol uses dual-funding, splicing and liquidity ads, leveraging
liquidity ads' extensions for paying funding fees.

It is recommended to use 0-conf, to avoid keeping upstream HTLCs held
for a long time and locking up liquidity in the network.