blip-0037: Bolt 12 encoded wallet `node_id`
[Bolt 12](https://github.com/lightning/bolts/pull/798) introduces a new
`sciddir_or_pubkey` field that can be used to identify the introduction
node of a blinded path. This is primarily used to reduce the size of
encoded offers, but this encoding trick can be extended to carry more
information about the node it's referencing. We introduce an additional
prefix to this field to tag `node_id`s that belong to mobile wallets.

When relaying a payment or an onion message, wallet providers usually
want to offer additional features to mobile wallets. Encoding the
information that the next node is a mobile wallet directly inside the
onion payload makes it easy for the relaying node to efficiently insert
additional steps before relaying (such as trying to wake-up the wallet
if it is offline) without negatively impacting relay to normal nodes.