blip-0034: the recommended_feerates message
Many lightning messages include feerates for on-chain transactions.
Nodes receiving those messages are supposed to reject them if they
are considered too low or too high. However, there is no mechanism
to let peers know which feerates we currently find acceptable, which
leads to frequent failures, especially when the mempool feerates
fluctuate quickly.

We introduce an optional message to advertise to our peers the feerates
we'd like to use when funding new or existing channels.