blip-0041: channel funding fee credit
Onboarding users to the lightning network is hard, because opening a
channel requires paying on-chain fees and allocating capital, hoping
that it will be used to earn routing fees. It is not economical nor
sustainable to open channels for users that receive a tiny amount and
may not use the network frequently enough.

Wallet providers usually work around this issue by requiring users to
receive a large enough first payment to pay on-chain fees for their
channel.

However, a collection of unrelated tiny payments received over a period
of time may be enough to justify creating a channel. We thus introduce
a funding fee credit, that lets users prepay for future channel(s) by
proactively revealing htlc preimages for small payments. When that fee
credit reaches a sufficient amount, it is used by the remote node to
cover the on-chain costs of opening a channel or splicing additional
liquidity.