Add bLIP 50: LSP Spec Transport Layer
The LSPS0 specification defines a protocol that is used for
communcations between LSP nodes and their clients. To this end it
utilizes a JSON-RPC format spoken over BOLT8 peer-to-peer messages.

The given protocol has been previously stabilized by the LSP
Spec group and is live in production with several LSP and client
implementations today.

As previously discussed on multiple occasions, the LSP Spec
group is however moving to a bLIP-centric process, which is why we
'upstream' these previously-stabilized specifications here.