Add bLIP 51: Channel Requests
The goal of LSPS1 is to provide a standardized LSP API based on
LSPS0 that allows clients to purchase a channel from an LSP
directly.

The given protocol has been previously stabilized by the LSP
Spec group and is live in production with several LSP and client
implementations today.

As previously discussed on multiple occasions, the LSP Spec
group is however moving to a bLIP-centric process, which is why we
'upstream' these previously-stabilized specifications here.