blip-0060: trampoline payments to legacy Bolt11 invoices
[Trampoline routing](https://github.com/lightning/bolts/pull/836) allows
nodes to send payments with limited access to the network graph, by
relying on trampoline nodes to compute the missing parts of the route.
The BOLTs only specify how such payments can be made to recipients who
either support the trampoline feature or use Bolt12 with a blinded path.
This leaves a gap when paying legacy nodes that don't support any of
those features.

This bLIP provides a mechanism to pay such legacy invoices, by revealing
some of the invoice data to the last trampoline node, who converts the
trampoline payment to a standard payment.