Add bLIP 55: Webhook Registration (LSPS5) (#55)
In typical mobile environments, a program that is not currently
being focused on by the user will be suspended, with all its
TCP connections dropped.

This specification provides a way for a mobile client to
register some specific webhook by which the LSP can signal a
push notification to the application developer server, which
will in turn convert the push notification to one that it itself
signs and can send to the mobile OS developer server.

The given protocol is based on LSPS0/bLIP 50 and has been previously
stabilized by the LSP Spec group.

As previously discussed on multiple occasions, the LSP Spec
group is however moving to a bLIP-centric process, which is why we
'upstream' these previously-stabilized specifications here.