bLIP-51: Specify invalid token error code
Previously the spec said that the LSP should return 'an error' in case
of an invalid or unrecogized token. To keep the approach similar to what
we do in bLIP-52 (where we specify a particular error code for this
case), we now here specify an error code to be used in this case.